+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to the "Project Structure Extractor" extension will be docum
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.0] - 2023-01-04
+### Added
+- Added a new configuration option `projectStructureExporter.showSize` to show file sizes in the project structure.
+- Added a new configuration option `projectStructureExporter.selectAndExtractButton` to enable a status bar button to extract the structure instead using command to extract the structure.
+- Status bar button "✨ Extract Structure" to select a folder and extract its structure to the clipboard.
+
+
 ## [0.0.3] - 2024-12-29
 
 ### Added
@@ -45,3 +52,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [0.0.1]: https://github.com/subucodes/vscode-ext-project-structure-extractor/releases/tag/v0.0.1
 [0.0.2]: https://github.com/subucodes/vscode-ext-project-structure-extractor/releases/tag/v0.0.2
 [0.0.3]: https://github.com/subucodes/vscode-ext-project-structure-extractor/releases/tag/v0.0.3
+[0.1.0]: https://github.com/subucodes/vscode-ext-project-structure-extractor/releases/tag/v0.1.0